Split the unittests for trainloop and result packages into multiple files for better readability
